+++ src/ESPRMBase.ts
@@ -24,7 +24,11 @@ import { ESPProvisionAdapterInterface } from "./types/provision";
 import { ESPTransportMode } from "./types/transport";
 import { ESPLocalDiscoveryAdapterInterface } from "./services/ESPTransport/ESPLocalDiscoveryAdapterInterface";
 import { ESPLocalControlAdapterInterface } from "./services/ESPTransport/ESPLocalControlAdapterInterface";
-import { ESPNotificationAdapterInterface } from "./types/adapter";
+import {
+  ESPNotificationAdapterInterface,
+  ESPOauthAdapterInterface,
+  ESPAppUtilityAdapterInterface,
+} from "./types/adapter";
 import { ESPRMStorageAdapterInterface } from "./types/storage";
 import { isValidEnumValue } from "./services/ESPRMHelpers/IsValidEnumValue";
 
@@ -68,6 +72,16 @@ export class ESPRMBase {
   static ESPNotificationAdapter: ESPNotificationAdapterInterface;
 
   /**
+   * Oauth adapter for the SDK.
+   */
+  static ESPOauthAdapter: ESPOauthAdapterInterface;
+
+  /**
+   * App utility adapter for the SDK.
+   */
+  static ESPAppUtilityAdapter: ESPAppUtilityAdapterInterface;
+
+  /**
    * Priority queue of transport modes.
    */
   static transportOrder: ESPTransportMode[] = [ESPTransportMode.cloud];
@@ -100,6 +114,12 @@ export class ESPRMBase {
     }
     if (config.notificationAdapter) {
       ESPRMBase.ESPNotificationAdapter = config.notificationAdapter;
+    }
+    if (config.oauthAdapter) {
+      ESPRMBase.ESPOauthAdapter = config.oauthAdapter;
+    }
+    if (config.appUtilityAdapter) {
+      ESPRMBase.ESPAppUtilityAdapter = config.appUtilityAdapter;
     }
     const apiManagerConfig: ESPRMAPIManagerConfig = {
       baseUrl: config.baseUrl,
@@ -205,6 +225,26 @@ export class ESPRMBase {
     adapter: ESPNotificationAdapterInterface
   ): void {
     ESPRMBase.ESPNotificationAdapter = adapter;
+  }
+
+  /**
+   * Sets the oauth adapter for the SDK.
+   *
+   * @param adapter - The oauth adapter implementing ESPOauthAdapterInterface.
+   */
+  public static setOauthAdapter(adapter: ESPOauthAdapterInterface): void {
+    ESPRMBase.ESPOauthAdapter = adapter;
+  }
+
+  /**
+   * Sets the app utility adapter for the SDK.
+   *
+   * @param adapter - The app utility adapter implementing ESPAppUtilityAdapterInterface.
+   */
+  public static setAppUtilityAdapter(
+    adapter: ESPAppUtilityAdapterInterface
+  ): void {
+    ESPRMBase.ESPAppUtilityAdapter = adapter;
   }
 
   /**