+++ src/utils/error/Error.ts
@@ -10,3 +10,4 @@ export { ESPStorageAdapterError } from "./ESPStorageAdapterError";
 export { ESPAPICallValidationError } from "./ESPAPICallValidationError";
 export { ESPTokenError } from "./ESPTokenError";
 export { ESPProvError } from "./ESPProvError";
+export { ESPAppPermissionError } from "./ESPAppPermissionError";